Restructure imports sorted by 'distance' from file
