+++ src/assets/icons/ValidationError.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+
 import { Theme } from '../../types'
 import { useThemeColor } from '../../utils'
