+++ src/assets/icons/enturLogo.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import EnturWhite from './../logos/Enturlogo_white.svg'
-import EnturBlack from './../logos/Enturlogo_black.svg'
+import EnturWhite from '../logos/Enturlogo_white.svg'
+import EnturBlack from '../logos/Enturlogo_black.svg'
 
 function EnturLogo({ className, style, height }: Props): JSX.Element {
     const enturLogo = style === 'black' ? EnturBlack : EnturWhite
