+++ src/assets/icons/positionPin.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Pin from './../logos/pin.svg'
+import Pin from '../logos/pin.svg'
 
 function PositionPin({ className, size }: Props): JSX.Element | null {
     return <img src={Pin} width={size} height={size} className={className} />
