+++ src/components/LoginModal/LoginOptions/index.tsx
@@ -3,10 +3,11 @@ import React, { Dispatch, SetStateAction } from 'react'
 import { Heading3, Paragraph } from '@entur/typography'
 import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton, SecondaryButton } from '@entur/button'
+
 import { ModalType, LoginCase } from '..'
 
-import sikkerhetBom from './../../../assets/images/sikkerhet_bom.png'
-import retinaSikkerhetBom from './../../../assets/images/sikkerhet_bom@2x.png'
+import sikkerhetBom from '../../../assets/images/sikkerhet_bom.png'
+import retinaSikkerhetBom from '../../../assets/images/sikkerhet_bom@2x.png'
 
 interface Props {
     setModalType: Dispatch<SetStateAction<ModalType>>
