+++ src/containers/MyBoards/BoardCard/index.tsx
@@ -7,8 +7,8 @@ import { LinkIcon, ClockIcon } from '@entur/icons'
 import { ThemeDashbboardPreview } from '../../../assets/icons/ThemeDashboardPreview'
 import { persist } from '../../../settings/FirestoreStorage'
 import { Settings } from '../../../settings'
-import BoardOverflowMenu from './OverflowMenu/.'
 
+import BoardOverflowMenu from './OverflowMenu'
 import './styles.scss'
 
 const DAYS = ['søn', 'man', 'tir', 'ons', 'tor', 'fre', 'lør']
